MOVE dataverk into src/ folder to ensure tests are run against installed version
ADD MANIFEST.in to include VERSION file
REMOVE unused import in requirements.txt
