+++ requirements.txt
@@ -11,7 +11,6 @@ google-resumable-media==0.3.2
 googleapis-common-protos==1.5.8
 google-cloud-bigquery==1.9.0
 hvac==0.7.2
-importlib_resources==1.0.2
 kafka-python==1.4.4
 numpy==1.15.2
 pandas==0.23.3
