+++ src/dataverk/utils/nbstripout.py
